Revert using IWorkbenchListener for cleanup on shutdown back to runtime shutdown hook
The original idea behind the switch was: Since KNIME was taking about a minute to entirely shutdown, I figured the VM was waiting
on the output/error reader threads of the processes, which would only terminate when the process they were reading on terminated
via timeout. The massive slowdown was caused by the synchronized issue fixed in one of the last commits, which makes this change
unnecessary and only caused "Workbench not created yet" exceptions.