Fix R connections leaking their workspaces
On windows Rserve does not automatically create a new workspace for every connection (which on
unix is automatically done because the process forks itself for every connection), so we need to clean
the workspace when releasing a connection. This could have been done via RConnection.detach(), but
that has a bug where we then cannot reconnect to the Rserve process, which means we have to clean
the workspace manually.