AP-6910: Fix race condition while opening R Snippet dialog
While opening the R Snippet conifgure dialog there was a chance that
a ArrayIndexOutOfBoundsException was thrown, preventing the dialog
from being opened.

Source of this bug was a race condition which occured while setting
the progress bar to a non-deterministic value
(m_processBar.setDeterministic(false)) which removes a hierarchy
listener from the list of listeners. Concurrently, the main thread
could be iterating over exactly this list.

Running the code which leads to the change of the progress bar in the
EDT resolves this problem.

AP-6910 (R Snippet Dialog does not open on first click on configure)