AP-6718: Do not transfer result of script twice
In some cases knime.out may be the last value returned by the script
result. Previously default was to resolve the reference to that result,
transferring the entire data (which would then later be fetched via
importBufferedDataTable("knime.out", ...) again).

We instead explicitly do not resolve the results of R scripts to save
performance.

A boolean parameter is added: eval(String, boolean), which signals
whether to resolve the result. This parameter can be used at various
places where the result of the script is not required.

AP-6718 (R nodes to send/receive large table chunk-wise (enable
data.frames >= 3GB))