AP-7410: Restore factor level order from 3.3
Since the changes for AP-6718 (transferring data chunk-wise) the factors
were created by R which orderes factor levels alphanumerically, but is
a deviation from the previous behaviour, where the factors were created
before transfer and in order of appearance. Since there may be workflows
which expect a certain level value to be first (for reference when training
models for example), we decided to bring back the original behaviour.

AP-7410: Failing tests