AP-7831: Fix race condition which causes potential loss of batch
A mechanism in the code which transfers the output data from R to KNIME
ensures that the last batch of rows is never smaller than 10k rows out of
performance reasons. If this happens, a Array needs to be resized.
Adding rows to the output container is done in parallel while receiving
data from R server via sockets and converting the R data to KNIME
DataCells. As waiting for "add rows to container" is done after resizing
of the array, there is a potential race condition in which the "add rows
to container" thread suddenly reads from the newly initialized array which
is filled with `null` and crashes. The batch is then lost.

AP-7831: R Integration: Race condition causes potential loss of rows in R output