Allow RHome path field to be wider
Because the label of the Rserve receiving buffer size is so large, the
RHome path field (which usually has wide content), will get squashed.
By splitting the label of the buffer size with a newline, the label
column in the preference page is significantly less wide, which in turn
the RHome field is able to use instead.

Signed-off-by: Jonathan Hale <jonathan.hale@knime.com>