AP-8333: Avoid resolving class names to class when loading templates
Which fails when the class lies outside of a bundle, like in the case
with RunRInMSSQLNodeFactory, as org.knime.r does not have access to it
through the bundle class loading.

To acchieve this, the template meta category has been changed from Class<?>
to String (the class name). This avoids the conversions in saving and
loading without changing any of the behaviour (apart from that).

Guessing the intended way to do this was to implement a TemplateProvider
and register it to the "org.knime.r.templaterepository" extension point,
which allows org.knime.r access to the classes of the bundle that wants
to add templates and therefore will make the class loadable when loading
the settings.

I find this rather overcomplicated, however, but maybe due to not seeing
the reason why the template meta category has to be a class.

AP-8333 (Fix issues with Run R Model on SQL Server)