AP-10947: Enhance Rserve installation instructions, to fix MacOS issues.
The latest version of Rserve, which is required to make the R nodes work
properly on mac is not available with the default package installation
method, users need to point the install to rforge. This is now reflected
in the error message.

AP-10947 (R version detection reports invalid installation when Rserve
(or any other dependency) is missing)