AP-12478: Wait for the stream reader threads
While retrieving R properties two threads are started to observe the
stdout and stderr of the R process. Sometimes the streams for the stdout
and stderr get closed while the threads are still reading the output
(not finished) which causes them to throw an IOException and log an
ERROR message.
Waiting for the threads before closing the stream solves this issue.