AP-14180: Added ability to set custom R version to be installed
The user has to confirm if she wants to install the default R version
set in the script or not. If 'no' is chosen, then the user is prompted
for the version she desires to install.

AP-14180 (Update R installation Script (Mac) to allow users to choose
which R version to be installed)