AP-18931: Do not use --vanilla when starting R or it won't find installed libraries
Starting with R 4.2.0 on Ubuntu the library search path is configured
via an environment variable in the default Renviron.site file. --vanilla
causes this file to not be considered. To consider this file in general
we replace the --vanilla flag with the flags --no-save --no-restore
--no-site-file and --no-init-file.

AP-18931 (R integration does not work with R 4.2.0 any more (on Ubuntu))